# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar alunos em uma escola.",
        "title": "Escola CRUD API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/alunos": {
            "get": {
                "description": "Retorna todos os alunos cadastrados no banco de dados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alunos"
                ],
                "summary": "Lista todos os alunos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Aluno"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona um novo aluno ao banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alunos"
                ],
                "summary": "Cria um novo aluno",
                "parameters": [
                    {
                        "description": "Dados do aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Aluno criado com sucesso!",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro na criação do aluno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/alunos/{id}": {
            "put": {
                "description": "Atualiza as informações de um aluno específico no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alunos"
                ],
                "summary": "Atualiza um aluno existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do aluno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do aluno",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aluno atualizado com sucesso!",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro na atualização do aluno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Exclui um aluno específico do banco de dados",
                "tags": [
                    "alunos"
                ],
                "summary": "Remove um aluno",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do aluno",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aluno removido com sucesso!",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro na remoção do aluno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Aluno": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "idade": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "nomeProfessor": {
                    "type": "string"
                },
                "notaPrimeiroSemestre": {
                    "type": "number"
                },
                "notaSegundoSemestre": {
                    "type": "number"
                },
                "numeroSala": {
                    "type": "integer"
                }
            }
        }
    }
}